+++ 4-full-mvc-framework/models/note_model.php
@@ -1,5 +1,9 @@
 <?php
 
+/**
+ * This is basically a simple CRUD demonstration.
+ */
+
 class Note_Model extends Model
 {
     public $errors = array();
@@ -46,7 +50,7 @@ class Note_Model extends Model
         if ($count == 1) {
             return true;
         } else {
-            $this->errors[] = "Database action was not successful.";
+            $this->errors[] = FEEDBACK_NOTE_CREATION_FAILED;
             return false;
         }
     }
@@ -66,7 +70,7 @@ class Note_Model extends Model
         if ($count == 1) {
             return true;
         } else {
-            $this->errors[] = "Database action was not successful.";
+            $this->errors[] = FEEDBACK_NOTE_EDITING_FAILED;
             return false;
         }                
                 
@@ -86,7 +90,7 @@ class Note_Model extends Model
         if ($count == 1) {
             return true;
         } else {
-            $this->errors[] = "Database action was not successful.";
+            $this->errors[] = FEEDBACK_NOTE_DELETION_FAILED;
             return false;
         }     
     }