+++ CHANGELOG.md
@@ -4,6 +4,7 @@ August 13th 2013
 * 4-full-mvc-framework: email change now asks for password (to prevent account takeovers)
 * 4-full-mvc-framework: "my account" menu item now links directly to user's profile
 * 4-full-mvc-framework: users can upgrade/downgrade their account status, like standard/premium etc.
+* 4-full-mvc-framework: system feedback messages are now constants, not pure text in the model anymore
 
 August 12th 2013
 * introducing the styles for the php-login project: https://github.com/panique/php-login-styles